git ignore add

+++ routes/board/delete.js
@@ -14,18 +14,20 @@ exports.delete_board = (req, res) => {
           }).catch(err => {
               res.status(500).json({ success: false });
           });
-          const { user_pk } = board;
-          const { pk : decoded_pk } = decoded;
-          if (decoded_pk == user_pk) {
-              await Board.destroy({
-                  where: { pk }
-              })
-              .catch(err => {
-                  res.status(500).json({ success: false });
-              });
-              res.status(200).json({ success: true });
-          } else{
-              res.status(412).json({ success: false });
+          if(board){
+            const { user_pk } = board;
+            const { pk : decoded_pk } = decoded;
+            if (decoded_pk == user_pk) {
+                await Board.destroy({
+                    where: { pk }
+                })
+                .catch(err => {
+                    res.status(500).json({ success: false });
+                });
+                res.status(200).json({ success: true });
+            } else{
+                res.status(412).json({ success: false });
+            }
           }
       } else{
         res.status(412).json({ success: false });
