+++ routes/boardComment/delete.js
@@ -15,18 +15,20 @@ exports.delete_comment = (req, res) => {
           .catch(err => {
             res.status(500).json({ success: false });
           });
-          const { user_pk } = comment;
-          const { pk : decoded_pk } = decoded;
-          if (decoded_pk == user_pk) {
-              await Comment.destroy({
-                  where: { pk }
-              })
-              .catch(err => {
-                  res.status(500).json({ success: false });
-              });
-              res.status(200).json({ success: true });
-          } else {
-              res.status(412).json({ success: false });
+          if(comment){
+            const { user_pk } = comment;
+            const { pk : decoded_pk } = decoded;
+            if (decoded_pk == user_pk) {
+                await Comment.destroy({
+                    where: { pk }
+                })
+                .catch(err => {
+                    res.status(500).json({ success: false });
+                });
+                res.status(200).json({ success: true });
+            } else {
+                res.status(412).json({ success: false });
+            }
           }
       } else {
         res.status(412).json({ success: false });
